+++ be/src/main/java/kr/codesquad/jazzmeet/review/entity/Review.java
@@ -1,4 +1,4 @@
-package kr.codesquad.jazzmeet.venue.entity;
+package kr.codesquad.jazzmeet.review.entity;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -7,6 +7,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import kr.codesquad.jazzmeet.global.time.BaseTimeEntity;
+import kr.codesquad.jazzmeet.venue.entity.Venue;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
