+++ be/src/main/java/kr/codesquad/jazzmeet/review/repository/ReviewRepository.java
@@ -1,8 +1,8 @@
-package kr.codesquad.jazzmeet.venue.repository;
+package kr.codesquad.jazzmeet.review.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kr.codesquad.jazzmeet.venue.entity.Review;
+import kr.codesquad.jazzmeet.review.entity.Review;
 
 public interface ReviewRepository extends JpaRepository<Review, Long> {
 }
