+++ be/src/main/java/kr/codesquad/jazzmeet/review/service/ReviewFacade.java
@@ -1,14 +1,15 @@
-package kr.codesquad.jazzmeet.venue.service;
+package kr.codesquad.jazzmeet.review.service;
 
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import kr.codesquad.jazzmeet.venue.dto.request.ReviewCreateRequest;
-import kr.codesquad.jazzmeet.venue.dto.response.ReviewCreateResponse;
-import kr.codesquad.jazzmeet.venue.entity.Review;
+import kr.codesquad.jazzmeet.review.dto.request.ReviewCreateRequest;
+import kr.codesquad.jazzmeet.review.dto.response.ReviewCreateResponse;
+import kr.codesquad.jazzmeet.review.entity.Review;
 import kr.codesquad.jazzmeet.venue.entity.Venue;
-import kr.codesquad.jazzmeet.venue.mapper.ReviewMapper;
+import kr.codesquad.jazzmeet.review.mapper.ReviewMapper;
+import kr.codesquad.jazzmeet.venue.service.VenueService;
 import lombok.RequiredArgsConstructor;
 
 @Transactional(readOnly = true)
